Create a PanicButton capsule.
